requirements.txt
PyAudio
SpeechRecognition
google-cloud-texttospeech==2.0.0
openai
